🧬 Rename the sub-package as *add*.
