+++ pymock_api/command/add/component.py
