+++ scripts/ci/get-unit-test-paths.sh
@@ -10,7 +10,7 @@ declare -a server_tests
 declare -a server_sgi_tests
 declare -a command_tests
 declare -a subcommand_run_tests
-declare -a subcommand_config_tests
+declare -a subcommand_add_tests
 declare -a subcommand_check_tests
 
 getalltests() {
@@ -48,11 +48,11 @@ getalltests() {
         # shellcheck disable=SC2124
         # shellcheck disable=SC2178
         subcommand_run_tests=${alltestpaths[@]}
-    elif echo "$1" | grep -q "config";  # /command/config
+    elif echo "$1" | grep -q "add";  # /command/add
     then
         # shellcheck disable=SC2124
         # shellcheck disable=SC2178
-        subcommand_config_tests=${alltestpaths[@]}
+        subcommand_add_tests=${alltestpaths[@]}
     elif echo "$1" | grep -q "check";  # /command/check
     then
         # shellcheck disable=SC2124
@@ -77,7 +77,7 @@ server_path=./test/unit_test/server/
 server_sgi_path=./test/unit_test/server/sgi/
 command_path=./test/unit_test/command/
 subcommand_run_path=./test/unit_test/command/run/
-subcommand_config_path=./test/unit_test/command/config/
+subcommand_add_path=./test/unit_test/command/add/
 subcommand_check_path=./test/unit_test/command/check/
 
 getalltests $init_path
@@ -87,10 +87,10 @@ getalltests $server_path
 getalltests $server_sgi_path
 getalltests $command_path
 getalltests $subcommand_run_path
-getalltests $subcommand_config_path
+getalltests $subcommand_add_path
 getalltests $subcommand_check_path
 
-dest=( "${init_tests[@]} ${utils_tests[@]} ${model_tests[@]} ${server_tests[@]} ${server_sgi_tests[@]} ${command_tests[@]} ${subcommand_run_tests[@]} ${subcommand_config_tests[@]} ${subcommand_check_tests[@]}" )
+dest=( "${init_tests[@]} ${utils_tests[@]} ${model_tests[@]} ${server_tests[@]} ${server_sgi_tests[@]} ${command_tests[@]} ${subcommand_run_tests[@]} ${subcommand_add_tests[@]} ${subcommand_check_tests[@]}" )
 
 if echo "$runtime_os" | grep -q "windows";
 then
